requirements.txt
# Minimal base environment for this repository's pipeline scripts
# (model inference, version/vuln analysis, dataset build, isolated eval runner).
loguru==0.7.3
openai==2.24.0
stdlib-list==0.12.0
requests==2.32.5
datasets==4.6.1
pyarrow==23.0.1
tqdm==4.67.3
python-dotenv==1.2.1
griffe[pypi]==2.0.1
matplotlib==3.10.0
tiktoken==0.12.0